[code beautify]1224 删除节点 删除连接
