+++ InternetGraph/GraphMatrix.h
@@ -7,14 +7,14 @@
 #include "vector"
 using namespace std;
 
-class GraphMatrix {
+class GraphMatrix {//矩阵
     public:
         GraphMatrix(int nodenum) {
             //初始化矩阵
             NodeNum = nodenum;
             M.resize(NodeNum);
-            for(int i=0; i<NodeNum; i++) {
-                M[i].resize(NodeNum);
+            for(int e=0; e<NodeNum; e++) {
+                M[e].resize(NodeNum);
             }
             //矩阵填充空边
 
@@ -23,13 +23,25 @@ class GraphMatrix {
                     GraphEdge EmptyEdge = GraphEdge();
                     M[i][j] = EmptyEdge;
                 }
-
-
         }
 
         void InsertNode(string Name) {
             GraphNode node = GraphNode(Name);
             Nodes.push_back(node);
+        }
+
+        void DeleteNode(string Name) {//删除节点函数
+            GraphNode node = SearchNodeName(Name);
+            if(node.getNodeName()!="Empty") {
+                cout<<"节点查找成功!"<<endl;
+                int idd=GetNodeId(Name);
+                cout<<GetNodeId(node.getNodeName())<<endl;
+                cout<<"删除节点..."<<endl;
+                Nodes.erase(Nodes.begin()+idd);
+
+            } else {
+                cout<<"未找到该节点"<<endl;
+            }
         }
 
         void InsertNode(string Name,string IP) {
@@ -71,9 +83,9 @@ class GraphMatrix {
         void SetConnect() {
             string node1name,node2name;
             cout<<"建立连接..."<<endl;
-            cout<<"输入节点1名称：";
+            cout<<"输入起始节点名称：";
             cin>>node1name;
-            cout<<"输入节点2名称：";
+            cout<<"输入终点节点名称：";
             cin>>node2name;
 
             GraphNode node1 = SearchNodeName(node1name);
@@ -91,23 +103,50 @@ class GraphMatrix {
                 //cout<<"Node:"<<node1ID<<"&&Node:"<<node2ID<<endl;
                 if(node2ID!=-1&&node1ID!=-1) {
                     M[node1ID][node2ID] = Link;
+                    M[node2ID][node1ID] = Link;
                 }
 
             }
-
-
         }
+
+        void DeleteConnect() {//删除边的函数
+            string node1name,node2name;
+            cout<<"删除连接..."<<endl;
+            cout<<"输入起始节点名称：";
+            cin>>node1name;
+            cout<<"输入终点节点名称：";
+            cin>>node2name;
+
+            GraphNode node1 = SearchNodeName(node1name);
+            GraphNode node2 = SearchNodeName(node2name);
+
+            if(node1.getNodeName()!="Empty"&&node2.getNodeName()!="Empty") {
+                cout<<"节点查找成功!"<<endl;
+                cout<<"删除连接..."<<endl;
+
+                //建立边
+                GraphEdge Link = GraphEdge(node1,node2,0);
+                //关联矩阵
+                int node1ID = GetNodeId(node1name);
+                int node2ID = GetNodeId(node2name);
+                //cout<<"Node:"<<node1ID<<"&&Node:"<<node2ID<<endl;
+                if(node2ID!=-1&&node1ID!=-1) {
+                    M[node1ID][node2ID] = Link;
+                    M[node2ID][node1ID] = Link;
+                }
+
+            }
+        }
+
 
         void DisplayMatrix() {
             cout<<setw(4)<<" ";
-            for(int k=0; k<M.size(); k++)
+            for(int k=0; k<Nodes.size(); k++)
                 cout<<setw(4)<<"N"<<k;
             cout<<endl;
-            for(int i=0; i<M.size(); i++) {
-				cout<<setw(4)<<"N"<<i;
-                for(int j=0; j<M[i].size(); j++) {
-                    
-
+            for(int i=0; i<Nodes.size(); i++) {
+                cout<<setw(4)<<"N"<<i;
+                for(int j=0; j<Nodes.size(); j++) {
                     if(M[i][j].getConnect())
                         cout<<setw(4)<<"1";
                     else if(!M[i][j].getConnect())
@@ -123,3 +162,4 @@ class GraphMatrix {
         vector<GraphNode> Nodes;		//存储节点数组
 };
 #endif
+
